# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Multi-tenant B2B SaaS Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface LoginInput {
  email: string;
  /** @minLength 6 */
  password: string;
}

export type AuthUserRole = typeof AuthUserRole[keyof typeof AuthUserRole];


export const AuthUserRole = {
  super_admin: 'super_admin',
  company_admin: 'company_admin',
  employee: 'employee',
} as const;

export interface AuthUser {
  id: number;
  email: string;
  name: string;
  role: AuthUserRole;
  /** @nullable */
  companyId?: number | null;
  /** @nullable */
  companyName?: string | null;
}

export interface Company {
  id: number;
  name: string;
  /** @nullable */
  taxId?: string | null;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  contactPhone?: string | null;
  /** @nullable */
  address?: string | null;
  employeeCount?: number;
  isActive?: boolean;
  createdAt: string;
}

export interface CompanyInput {
  /** @minLength 1 */
  name: string;
  taxId?: string;
  contactEmail?: string;
  contactPhone?: string;
  address?: string;
}

export interface CompanyUpdate {
  /** @minLength 1 */
  name?: string;
  taxId?: string;
  contactEmail?: string;
  contactPhone?: string;
  address?: string;
  isActive?: boolean;
}

export interface CompanyListResponse {
  data: Company[];
  total: number;
  page: number;
  limit: number;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  super_admin: 'super_admin',
  company_admin: 'company_admin',
  employee: 'employee',
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  /** @nullable */
  companyId?: number | null;
  /** @nullable */
  companyName?: string | null;
  isActive?: boolean;
  createdAt: string;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  company_admin: 'company_admin',
  employee: 'employee',
} as const;

export interface UserInput {
  email: string;
  /** @minLength 1 */
  name: string;
  role: UserInputRole;
  /** @minLength 6 */
  password: string;
  companyId?: number;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  company_admin: 'company_admin',
  employee: 'employee',
} as const;

export interface UserUpdate {
  /** @minLength 1 */
  name?: string;
  role?: UserUpdateRole;
  isActive?: boolean;
  /** @minLength 6 */
  password?: string;
}

export interface UserListResponse {
  data: User[];
  total: number;
  page: number;
  limit: number;
}

export interface Employee {
  id: number;
  companyId: number;
  /** @nullable */
  nationalId?: string | null;
  /** @nullable */
  registrationNo?: string | null;
  firstName: string;
  lastName: string;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  position?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  hireDate?: string | null;
  isActive?: boolean;
  createdAt: string;
}

export interface EmployeeInput {
  companyId?: number;
  nationalId?: string;
  registrationNo?: string;
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  department?: string;
  position?: string;
  email?: string;
  phone?: string;
  hireDate?: string;
}

export interface EmployeeUpdate {
  nationalId?: string;
  registrationNo?: string;
  /** @minLength 1 */
  firstName?: string;
  /** @minLength 1 */
  lastName?: string;
  department?: string;
  position?: string;
  email?: string;
  phone?: string;
  hireDate?: string;
  isActive?: boolean;
}

export interface EmployeeListResponse {
  data: Employee[];
  total: number;
  page: number;
  limit: number;
}

export interface EmployeeUploadInput {
  companyId: number;
}

export interface UploadError {
  row: number;
  message: string;
  rawData?: string;
}

export interface EmployeeUploadResult {
  imported: number;
  failed: number;
  errors: UploadError[];
}

export interface DashboardStats {
  totalEmployees: number;
  totalCompanies: number;
  totalUsers: number;
  activeEmployees: number;
  totalUploads: number;
  newEmployeesThisMonth?: number;
}

export interface DepartmentStat {
  department: string;
  count: number;
}

export interface UploadLog {
  id: number;
  companyId: number;
  companyName: string;
  importedCount: number;
  failedCount: number;
  uploadedAt: string;
  uploadedBy: string;
}

export type ListCompaniesParams = {
page?: number;
limit?: number;
search?: string;
};

export type ListUsersParams = {
page?: number;
limit?: number;
companyId?: number;
};

export type ListEmployeesParams = {
page?: number;
limit?: number;
search?: string;
department?: string;
companyId?: number;
};

export type GetRecentUploadsParams = {
limit?: number;
};