# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Multi-tenant B2B SaaS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List all companies (Super Admin only)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyListResponse"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Create a new company (Super Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInput"
      responses:
        "201":
          description: Company created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{id}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get a company by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Company details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCompany
      tags: [companies]
      summary: Update a company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUpdate"
      responses:
        "200":
          description: Company updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    delete:
      operationId: deleteCompany
      tags: [companies]
      summary: Delete a company (Super Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Company deleted
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List employees (scoped to company)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: department
          in: query
          schema:
            type: string
        - name: companyId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeListResponse"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create a new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Employee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get a single employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Employee updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Employee deleted
  /employees/upload:
    post:
      operationId: uploadEmployees
      tags: [employees]
      summary: Upload an Excel file to bulk-import employees
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/EmployeeUploadInput"
      responses:
        "200":
          description: Upload result with success and error rows
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeUploadResult"
        "400":
          description: Bad file or parse error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 6
    AuthUser:
      type: object
      required: [id, email, name, role]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [super_admin, company_admin, employee]
        companyId:
          type: ["integer", "null"]
        companyName:
          type: ["string", "null"]

    # Company schemas
    Company:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        taxId:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        employeeCount:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CompanyInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        taxId:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        address:
          type: string
    CompanyUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        taxId:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        address:
          type: string
        isActive:
          type: boolean
    CompanyListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Company"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # User schemas
...
    Employee:
      type: object
      required: [id, companyId, firstName, lastName, createdAt]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        nationalId:
          type: ["string", "null"]
        registrationNo:
          type: ["string", "null"]
        firstName:
          type: string
        lastName:
          type: string
        department:
          type: ["string", "null"]
        position:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        hireDate:
          type: ["string", "null"]
          format: date
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    EmployeeInput:
      type: object
      required: [firstName, lastName]
      properties:
        companyId:
          type: integer
        nationalId:
          type: string
        registrationNo:
          type: string
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        department:
          type: string
        position:
          type: string
        email:
          type: string
        phone:
          type: string
        hireDate:
          type: string
    EmployeeUpdate:
      type: object
      properties:
        nationalId:
          type: string
        registrationNo:
          type: string
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        department:
          type: string
        position:
          type: string
        email:
          type: string
        phone:
          type: string
        hireDate:
          type: string
        isActive:
          type: boolean
    EmployeeListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Employee"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    EmployeeUploadInput:
      type: object
      required: [companyId]
      properties:
        companyId:
          type: integer
    EmployeeUploadResult:
      type: object
      required: [imported, failed, errors]
      properties:
        imported:
          type: integer
        failed:
          type: integer
        errors:
          type: array
          items:
            $ref: "#/components/schemas/UploadError"
    UploadError:
      type: object
      required: [row, message]
      properties:
        row:
          type: integer
        message:
          type: string
        rawData:
          type: string

    # Dashboard schemas
...