# lib/api-zod/src/generated/types/company.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Multi-tenant B2B SaaS Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Company {
  id: number;
  name: string;
  /** @nullable */
  taxId?: string | null;
  /** @nullable */
  contactEmail?: string | null;
  /** @nullable */
  contactPhone?: string | null;
  /** @nullable */
  address?: string | null;
  employeeCount?: number;
  isActive?: boolean;
  createdAt: Date;
}