# lib/api-zod/src/generated/types/companyListResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Multi-tenant B2B SaaS Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Company } from './company';

export interface CompanyListResponse {
  data: Company[];
  total: number;
  page: number;
  limit: number;
}