# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Multi-tenant B2B SaaS Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Employee {
  id: number;
  companyId: number;
  /** @nullable */
  nationalId?: string | null;
  /** @nullable */
  registrationNo?: string | null;
  firstName: string;
  lastName: string;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  position?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  hireDate?: Date | null;
  isActive?: boolean;
  createdAt: Date;
}