# lib/api-zod/src/generated/types/employeeInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Multi-tenant B2B SaaS Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface EmployeeInput {
  companyId?: number;
  nationalId?: string;
  registrationNo?: string;
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  department?: string;
  position?: string;
  email?: string;
  phone?: string;
  hireDate?: string;
}